requirements.txt
python-jenkins
click
